flambda-backend: Use `Compilation_unit.t` instead of `Ident.t` for globals
Instead of a `getglobal` primitive taking an `Ident.t`, Lambda
now has a `getglobal` taking a `Compilation_unit.t` and a
separate `getpredef` taking a predef `Ident.t`. (It would be
sensible to define a new type `Predef.t` or something.)

In order to generate the full `Compilation_unit.t` when producing
Lambda, we need to know the pack prefix of each global used in the
source file. In particular, it means we need the pack prefix in the
`.cmi` format. This suffices even with `-no-alias-deps` (that is,
when we might be missing the relevant `.cmi`), since a module that's
merely aliased never produces a `getglobal` operation.

In return, packing no longer requires fiddling with the `Ident.t`s in
the code of the component modules - they are created correctly to
begin with.

These changes don't reach much into the bytecode compiler or the
middle ends: the `Compilation_unit.t` gets translated into an
`Ident.t` or `Symbol.t`, respectively, on the way in. Many fiddly
details at the driver level are affected, however, since we've been
treating `modname` and `Ident.t` as interchangeable.

Finally, I've removed the `Misc.modname` type (which was simply an
alias to `string`) in favor of `Compilation_unit.t` or
`Compilation_unit.Name.t`, as applicable. (Indeed, making the
difference explicit is a significant improvement.)