flambda-backend: Keep original naming for globals in bytecode
This branch was previously using the same name mangling to convert compilation
units into linkage names (for asm) and global idents (for bytecode). That was
unnecessary. This returns to the original behavior and improves the API for
getting idents out of compilation units to emphasize that it's only for use in
generating bytecode. (It also removes the one place outside of bytecode
generation that we did the CU-to-ident conversion, which was wholly
unnecessary.)