flambda-backend: Allow a pack to access its own members and re-enable check
I had written a check in `Persistent_env` to make sure nothing can directly
access the members of other packs, but disabled it because it broke packing.
Having realized that the check simply needed to be loosened to allow a pack
to access its own members, I've now re-enabled it.