flambda-backend: Changes to representations of import lists
An interface import is now a `CU.Name.t * (CU.t * Digest.t) option`, reflecting
that if there is a .cmi available, we will know the full path of the compilation
unit.

An implementation import is now a `CU.t * Digest.t option`, reflecting that
there will always be a .cmi available, so we'll always know the full path, but
there may not be a .cmx available.